requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langgraph>=0.2.69
langchain>=0.1.11
langchain-openai>=0.1.4
langsmith>=0.3.12
python-dotenv>=1.0.1
matplotlib>=3.10.1
networkx>=3.4.2
pydantic>=2.6.4
pytest>=7.4.0
python-multipart>=0.0.6
typing-extensions>=4.8.0
openai>=1.0.0
requests>=2.31.0
uuid>=1.30
datetime>=5.3
jinja2>=3.1.2 